test_params_on_val/requirements.txt
numpy==1.18.5
pyclipper
colorlog
Polygon3
opencv-python 
torch
torchvision
tensorboard
matplotlib
scikit-image
tqdm
sklearn